Create Header.jsx
Every time the path changes, it will update the title of the page to include the path name. Contains a Navbar with links to all the available pages. For the portfolio, it contains a "NavDropdown" with all the projects as "NavDropdown.Item". When items are not loaded yet, it will show a Spinner.