Create ProjectCard.jsx
Display an image, a title, a description, and buttons to the repository and live demo of a project. Button to the next project or if it is the last project it will go to the contact page. If it is not the first project, it will also contain a button to the previous project.